Cleaned up the types for validation. Added a base IValidator interface that specifies the base interface for a regular or composite validator (validate or getErrorCount). Renamed the existing one to ISimpleValidator
